refactor: Separate diner image service from diner service
